# openapi.yaml
...
openapi: 3.2.1
info:
  title: Moseby
  description: Resort concierge API for stays, guests, rooms and activities.
  version: 0.1.0
paths:
  /hotels:
    get:
      summary: List Hotels
      operationId: listHotels
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            minLength: 1
            maxLength: 2048
          - type: 'null'
          description: Continue the same query.
          title: Cursor
        description: Continue the same query.
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_Hotel_'
        '501':
          description: Operation is not implemented.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-permissions:
        anyOf:
        - moseby.hotels:read
        - moseby:read
      x-hotel-scoped: true
  /hotels/{id}:
    get:
      summary: Get Hotel
      operationId: getHotel
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          pattern: ^hotel_[0-7][0-9A-HJKMNP-TV-Z]{25}$
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hotel'
        '501':
          description: Operation is not implemented.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-permissions:
        anyOf:
        - moseby.hotels:read
        - moseby:read
      x-hotel-scoped: true
  /staff-members:
    get:
      summary: List Staff Members
      operationId: listStaffMembers
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            minLength: 1
            maxLength: 2048
          - type: 'null'
          description: Continue the same query.
          title: Cursor
        description: Continue the same query.
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_StaffMember_'
        '501':
          description: Operation is not implemented.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-permissions:
        anyOf:
        - moseby.staff-members:read
        - moseby:read
      x-hotel-scoped: true
  /staff-members/{id}:
    get:
      summary: Get Staff Member
      operationId: getStaffMember
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          pattern: ^staff_member_[0-7][0-9A-HJKMNP-TV-Z]{25}$
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaffMember'
        '501':
          description: Operation is not implemented.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-permissions:
        anyOf:
        - moseby.staff-members:read
        - moseby:read
      x-hotel-scoped: true
  /venues:
    get:
      summary: List Venues
      operationId: listVenues
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            minLength: 1
            maxLength: 2048
          - type: 'null'
          description: Continue the same query.
          title: Cursor
        description: Continue the same query.
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_Venue_'
        '501':
          description: Operation is not implemented.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-permissions:
        anyOf:
        - moseby.venues:read
        - moseby:read
      x-hotel-scoped: false
  /venues/{id}:
    get:
      summary: Get Venue
      operationId: getVenue
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          pattern: ^venue_[0-7][0-9A-HJKMNP-TV-Z]{25}$
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Venue'
        '501':
          description: Operation is not implemented.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-permissions:
        anyOf:
        - moseby.venues:read
        - moseby:read
      x-hotel-scoped: false
  /parties/{id}:
    get:
      summary: Get Party
      operationId: getParty
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          pattern: ^party_[0-7][0-9A-HJKMNP-TV-Z]{25}$
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Party'
        '501':
          description: Operation is not implemented.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-permissions:
        anyOf:
        - moseby.guests:read
        - moseby:read
      x-hotel-scoped: true
...
  /party-details/{id}:
    get:
      summary: Get Party Detail
      operationId: getPartyDetail
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          pattern: ^party_detail_[0-7][0-9A-HJKMNP-TV-Z]{25}$
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyDetail'
        '501':
          description: Operation is not implemented.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-permissions:
        anyOf:
        - moseby.guests:read
        - moseby:read
      x-hotel-scoped: true
  /party-details:
    post:
      summary: Create Party Detail
      description: Record a note about a party and identify any guests mentioned.
      operationId: createPartyDetail
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePartyDetailRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyDetail'
        '501':
          description: Operation is not implemented.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-permissions:
        anyOf:
        - moseby.guests:write
        - moseby:write
      x-hotel-scoped: true
    query:
      summary: Search Party Details
      operationId: searchPartyDetails
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_PartyDetail_'
        '501':
          description: Operation is not implemented.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-permissions:
        anyOf:
        - moseby.guests:read
        - moseby:read
      x-hotel-scoped: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchPartyDetailsRequest'
  /room-keys:
    get:
      summary: List Room Keys
      operationId: listRoomKeys
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            minLength: 1
            maxLength: 2048
          - type: 'null'
          description: Continue the same query.
          title: Cursor
        description: Continue the same query.
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_RoomKey_'
        '501':
          description: Operation is not implemented.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-permissions:
        anyOf:
        - moseby.bookings:read
        - moseby:read
      x-hotel-scoped: true
    post:
      summary: Issue Room Key
      operationId: issueRoomKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueRoomKeyRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomKey'
        '501':
          description: Operation is not implemented.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-permissions:
        anyOf:
        - moseby.bookings:write
        - moseby:write
      x-hotel-scoped: true
  /room-keys/{id}:
    get:
      summary: Get Room Key
      operationId: getRoomKey
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          pattern: ^room_key_[0-7][0-9A-HJKMNP-TV-Z]{25}$
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomKey'
        '501':
          description: Operation is not implemented.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-permissions:
        anyOf:
        - moseby.bookings:read
        - moseby:read
      x-hotel-scoped: true
  /room-keys/{id}:deactivate:
    post:
      summary: Deactivate Room Key
      operationId: deactivateRoomKey
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          pattern: ^room_key_[0-7][0-9A-HJKMNP-TV-Z]{25}$
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeactivateRoomKeyRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomKey'
        '501':
          description: Operation is not implemented.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-permissions:
        anyOf:
        - moseby.bookings:write
        - moseby:write
      x-hotel-scoped: true
...
components:
  schemas:
...
    Address:
      properties:
        line_1:
          anyOf:
          - type: string
          - type: 'null'
          title: Line 1
        line_2:
          anyOf:
          - type: string
          - type: 'null'
          title: Line 2
        city:
          anyOf:
          - type: string
          - type: 'null'
          title: City
        postcode:
          anyOf:
          - type: string
          - type: 'null'
          title: Postcode
        country_code:
          anyOf:
          - type: string
            pattern: ^[A-Z]{2}$
          - type: 'null'
          title: Country Code
          description: Two-letter ISO 3166-1 country code.
      additionalProperties: false
      type: object
      title: Address
      description: Postal location; fields may be absent where they do not apply.
...
    BookingId:
      type: string
      pattern: ^booking_[0-7][0-9A-HJKMNP-TV-Z]{25}$
...
    CreatePartyDetailRequest:
      properties:
        request_id:
          type: string
          minLength: 1
          title: Request Id
          description: Identifies this logical request; reuse it when retrying.
        payload:
          $ref: '#/components/schemas/CreatePartyDetailRequestPayload'
      additionalProperties: false
      type: object
      required:
      - request_id
      - payload
      title: CreatePartyDetailRequest
    CreatePartyDetailRequestPayload:
      properties:
        party_id:
          $ref: '#/components/schemas/PartyId'
        text:
          type: string
          minLength: 1
          title: Text
          description: Record the evidence; the service resolves guest references.
      additionalProperties: false
      type: object
      required:
      - party_id
      - text
      title: CreatePartyDetailRequestPayload
...
    DeactivateRoomKeyRequest:
      properties:
        request_id:
          type: string
          minLength: 1
          title: Request Id
          description: Identifies this logical request; reuse it when retrying.
        payload:
          $ref: '#/components/schemas/DeactivateRoomKeyRequestPayload'
      additionalProperties: false
      type: object
      required:
      - request_id
      - payload
      title: DeactivateRoomKeyRequest
    DeactivateRoomKeyRequestPayload:
      properties:
        reason:
          type: string
          minLength: 1
          title: Reason
      additionalProperties: false
      type: object
      required:
      - reason
      title: DeactivateRoomKeyRequestPayload
...
    GuestId:
      type: string
      pattern: ^guest_[0-7][0-9A-HJKMNP-TV-Z]{25}$
    GuestReferenceStatus:
      type: string
      enum:
      - GUEST_REFERENCE_STATUS_PENDING
      - GUEST_REFERENCE_STATUS_RESOLVED
      - GUEST_REFERENCE_STATUS_AMBIGUOUS
      - GUEST_REFERENCE_STATUS_FAILED
      title: GuestReferenceStatus
      description: Ambiguous notes retain clear guest matches while other references
        remain uncertain.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    Hotel:
      properties:
        id:
          $ref: '#/components/schemas/HotelId'
        name:
          type: string
          minLength: 1
          title: Name
        address:
          $ref: '#/components/schemas/Address'
        created_at:
          type: string
          format: date-time
          title: Created At
      additionalProperties: false
      type: object
      required:
      - id
      - name
      - address
      - created_at
      title: Hotel
      description: The location owning a booking and its rooms.
    HotelId:
      type: string
      pattern: ^hotel_[0-7][0-9A-HJKMNP-TV-Z]{25}$
...
    IdFilter_GuestId_:
      items:
        $ref: '#/components/schemas/GuestId'
      type: array
      maxItems: 100
      minItems: 1
      description: Match any listed ID; duplicates do not multiply results.
...
    IdFilter_PartyId_:
      items:
        $ref: '#/components/schemas/PartyId'
      type: array
      maxItems: 100
      minItems: 1
      description: Match any listed ID; duplicates do not multiply results.
...
    IssueRoomKeyRequest:
      properties:
        request_id:
          type: string
          minLength: 1
          title: Request Id
          description: Identifies this logical request; reuse it when retrying.
        payload:
          $ref: '#/components/schemas/IssueRoomKeyRequestPayload'
      additionalProperties: false
      type: object
      required:
      - request_id
      - payload
      title: IssueRoomKeyRequest
    IssueRoomKeyRequestPayload:
      properties:
        room_reservation_id:
          $ref: '#/components/schemas/RoomReservationId'
        code:
          anyOf:
          - type: string
          - type: 'null'
          title: Code
      additionalProperties: false
      type: object
      required:
      - room_reservation_id
      title: IssueRoomKeyRequestPayload
...
    Page_Hotel_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/Hotel'
          type: array
          title: Items
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
          description: Opaque continuation token; null means there are no more results.
      additionalProperties: false
      type: object
      required:
      - items
      - next_cursor
      title: Page[Hotel]
...
    Page_PartyDetail_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/PartyDetail'
          type: array
          title: Items
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
          description: Opaque continuation token; null means there are no more results.
      additionalProperties: false
      type: object
      required:
      - items
      - next_cursor
      title: Page[PartyDetail]
...
    Page_RoomKey_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/RoomKey'
          type: array
          title: Items
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
          description: Opaque continuation token; null means there are no more results.
      additionalProperties: false
      type: object
      required:
      - items
      - next_cursor
      title: Page[RoomKey]
...
    Page_StaffMember_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/StaffMember'
          type: array
          title: Items
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
          description: Opaque continuation token; null means there are no more results.
      additionalProperties: false
      type: object
      required:
      - items
      - next_cursor
      title: Page[StaffMember]
...
    Page_Venue_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/Venue'
          type: array
          title: Items
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
          description: Opaque continuation token; null means there are no more results.
      additionalProperties: false
      type: object
      required:
      - items
      - next_cursor
      title: Page[Venue]
    Party:
      properties:
        id:
          $ref: '#/components/schemas/PartyId'
        booking_id:
          $ref: '#/components/schemas/BookingId'
        created_at:
          type: string
          format: date-time
          title: Created At
      additionalProperties: false
      type: object
      required:
      - id
      - booking_id
      - created_at
      title: Party
      description: The party for one confirmed booking; guests reference this party.
    PartyDetail:
      properties:
        id:
          $ref: '#/components/schemas/PartyDetailId'
        party_id:
          $ref: '#/components/schemas/PartyId'
        text:
          type: string
          minLength: 1
          title: Text
        referenced_guest_ids:
          items:
            $ref: '#/components/schemas/GuestId'
          type: array
          title: Referenced Guest Ids
          description: Clear matches from this party, validated by the service. An
            ambiguous status means other references remain unclear.
        reference_format_version:
          type: integer
          const: 1
          title: Reference Format Version
          description: Version of the stored reference format.
          default: 1
        reference_status:
          $ref: '#/components/schemas/GuestReferenceStatus'
        created_at:
          type: string
          format: date-time
          title: Created At
      additionalProperties: false
      type: object
      required:
      - id
      - party_id
      - text
      - referenced_guest_ids
      - reference_status
      - created_at
      title: PartyDetail
      description: Reported evidence about a party; inferred references do not update
        guest facts.
    PartyDetailId:
      type: string
      pattern: ^party_detail_[0-7][0-9A-HJKMNP-TV-Z]{25}$
    PartyId:
      type: string
      pattern: ^party_[0-7][0-9A-HJKMNP-TV-Z]{25}$
...
    RoomKey:
      properties:
        id:
          $ref: '#/components/schemas/RoomKeyId'
        room_reservation_id:
          $ref: '#/components/schemas/RoomReservationId'
        effective:
          type: boolean
          title: Effective
          description: 'Access now: not deactivated, with a confirmed booking and
            an uncancelled reservation whose current date range includes now.'
        code:
          anyOf:
          - type: string
          - type: 'null'
          title: Code
          description: Optional demo key label; not a live door credential.
        deactivated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deactivated At
          description: Explicit key revocation time; parent cancellation does not
            set this.
        deactivation_reason:
          anyOf:
          - type: string
            minLength: 1
          - type: 'null'
          title: Deactivation Reason
        created_at:
          type: string
          format: date-time
          title: Created At
      additionalProperties: false
      type: object
      required:
      - id
      - room_reservation_id
      - effective
      - code
      - deactivated_at
      - deactivation_reason
      - created_at
      title: RoomKey
      description: Anonymous access tied to one reservation; a later stay cannot revive
        this key.
    RoomKeyId:
      type: string
      pattern: ^room_key_[0-7][0-9A-HJKMNP-TV-Z]{25}$
...
    RoomReservationId:
      type: string
      pattern: ^room_reservation_[0-7][0-9A-HJKMNP-TV-Z]{25}$
...
    SearchPartyDetailsRequest:
      properties:
        request_id:
          type: string
          minLength: 1
          title: Request Id
          description: Identifies this logical request; reuse it when retrying.
        payload:
          $ref: '#/components/schemas/SearchPartyDetailsRequestPayload'
      additionalProperties: false
      type: object
      required:
      - request_id
      - payload
      title: SearchPartyDetailsRequest
    SearchPartyDetailsRequestPayload:
      properties:
        cursor:
          anyOf:
          - type: string
            maxLength: 2048
            minLength: 1
          - type: 'null'
          title: Cursor
          description: Continue the same query.
        limit:
          type: integer
          maximum: 100.0
          minimum: 1.0
          title: Limit
          default: 50
        party_ids:
          $ref: '#/components/schemas/IdFilter_PartyId_'
        guest_ids:
          anyOf:
          - $ref: '#/components/schemas/IdFilter_GuestId_'
          - type: 'null'
          description: Restrict to details referencing any listed guest.
        text:
          anyOf:
          - type: string
            minLength: 1
          - type: 'null'
          title: Text
          description: Whole-word keywords, all required in any order; case and Latin
            accents are ignored. Punctuation separates words.
      additionalProperties: false
      type: object
      required:
      - party_ids
      title: SearchPartyDetailsRequestPayload
...
    StaffMember:
      properties:
        id:
          $ref: '#/components/schemas/StaffMemberId'
        hotel_id:
          $ref: '#/components/schemas/HotelId'
        staff_code:
          type: string
          minLength: 1
          title: Staff Code
          description: Human-facing code, separate from the resource ID.
        first_name:
          type: string
          minLength: 1
          title: First Name
        last_name:
          type: string
          minLength: 1
          title: Last Name
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        role:
          $ref: '#/components/schemas/StaffRole'
        created_at:
          type: string
          format: date-time
          title: Created At
      additionalProperties: false
      type: object
      required:
      - id
      - hotel_id
      - staff_code
      - first_name
      - last_name
      - title
      - role
      - created_at
      title: StaffMember
      description: The acting staff identity carried through requests and jobs.
    StaffMemberId:
      type: string
      pattern: ^staff_member_[0-7][0-9A-HJKMNP-TV-Z]{25}$
    StaffRole:
      type: string
      enum:
      - STAFF_ROLE_UNKNOWN
      - STAFF_ROLE_CONCIERGE
      title: StaffRole
      description: Initial demo role; unknown grants no permissions.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    Venue:
      properties:
        id:
          $ref: '#/components/schemas/VenueId'
        name:
          type: string
          minLength: 1
          title: Name
        address:
          $ref: '#/components/schemas/Address'
        capacity:
          type: integer
          minimum: 0.0
          title: Capacity
          description: Administrative capacity; activities enforce their own limits.
        created_at:
          type: string
          format: date-time
          title: Created At
      additionalProperties: false
      type: object
      required:
      - id
      - name
      - address
      - capacity
      - created_at
      title: Venue
      description: A shared activity location without a hotel owner.
    VenueId:
      type: string
      pattern: ^venue_[0-7][0-9A-HJKMNP-TV-Z]{25}$